requirements.txt
numpy>=1.26
scikit-learn>=1.4
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
torch>=2.2.0
tqdm>=4.66.0
PyPdf2>=1.28.0
groq>=0.4.0
langchain-text-splitters>=0.0.1